Authenticate task endpoints

+++ 3_task-app/src/routers/task.js
@@ -14,20 +14,25 @@ router.post('/tasks', auth, async (req, res) => {
     }
 });
 
-router.get('/tasks', async (req, res) => {
+router.get('/tasks', auth, async (req, res) => {
     try {
-        const tasks = await Task.find({});
+        // Solution 1
+        const tasks = await Task.find({ owner: req.user._id });
         res.send(tasks);
+        // Solution 2
+        // await req.user.populate('task').execPopulate();
+        // res.send(req.user.tasks);
     } catch (error) {
         res.status(500).send(error);
     }
 });
 
-router.get('/tasks/:id', async (req, res) => {
+router.get('/tasks/:id', auth, async (req, res) => {
     const _id = req.params.id;
 
     try {
-        const task = await Task.findById(_id);
+        const task = await Task.findOne({ _id, owner: req.user._id });
+
         if (!task) return res.status(404).send();
         res.send(task);
     } catch (error) {
@@ -35,7 +40,7 @@ router.get('/tasks/:id', async (req, res) => {
     }
 });
 
-router.patch('/tasks/:id', async (req, res) => {
+router.patch('/tasks/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body);
     const allowedUpdates = ['description', 'completed'];
     const isValidOperation = updates.every(update => allowedUpdates.includes(update));
@@ -43,21 +48,21 @@ router.patch('/tasks/:id', async (req, res) => {
     if (!isValidOperation) return res.status(400), send({ error: 'Invalid updates.' });
 
     try {
-        const task = await Task.findById(req.params.id);
+        const task = await Task.findOne({ _id: req.params.id, owner: req.user._id });
+
+        if (!task) return res.status(404).send();
 
         updates.forEach(update => task[update] = req.body[update]);
         await task.save();
-
-        if (!task) return res.status(404).send();
         res.send(task);
     } catch (error) {
         res.status(400).send(error);
     }
 });
 
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
     try {
-        const task = await Task.findByIdAndDelete(req.params.id);
+        const task = await Task.findByIdAndDelete({ _id: req.params.id, owner: req.user._id });
 
         if (!task) return res.status(404).send();
         res.send(task);
